fix(react-router): fix non-portable type definitions
