+++ packages/react-router/src/@types/core.ts
@@ -1,2 +1,1 @@
-/** Re-exports `@aura-stack/react/types` for `@aura-stack/react-router` consumers. */
-export type * from "@aura-stack/react/types"
+export * from "@aura-stack/react/types"
