--- packages/react-router/src/client.ts
@@ -1,3 +1,0 @@
-export { createAuthClient, useAuthActions, useSession, useSignIn, useSignInCredentials, useSignOut } from "@aura-stack/react"
-export type { AuthClientOptions } from "@aura-stack/react"
-export { AuthProvider, type AuthProviderProps } from "@/context"
