+++ packages/react-router/src/client.tsx
@@ -1,0 +1,11 @@
+export {
+    createAuthClient,
+    useAuthActions,
+    useSession,
+    useSignIn,
+    useSignInCredentials,
+    useUpdateSession,
+    useSignOut,
+    type AuthClientOptions,
+} from "@aura-stack/react"
+export { AuthProvider, type AuthProviderProps } from "@/context"
