+++ packages/react-router/src/index.ts
@@ -1,2 +1,2 @@
 export { createAuth } from "@/createAuth"
-export type { User, Session, AuthConfig, AuthInstance } from "@aura-stack/react"
+export type { User, Session, AuthConfig, AuthInstance } from "@/@types/core"
