+++ packages/react-router/tsdown.config.ts
@@ -3,9 +3,10 @@ import { tsdownConfig } from "@aura-stack/tsdown-config"
 
 export default defineConfig({
     ...tsdownConfig,
+    platform: "neutral",
     entry: [
         "src/index.ts",
-        "src/client.ts",
+        "src/client.tsx",
         "src/oauth/index.ts",
         "src/_core/identity.ts",
         "src/_core/crypto.ts",
